vms-desktop-app/requirements.txt
numpy >= 1.16.0
# opencv-contrib-python >= 4.0.0.21
scikit-learn >= 0.20.2
# for DL
tensorflow >= 1.12.0
keras >= 2.2.4
# VGGface
git+https://github.com/rcmalli/keras-vggface.git
# for cryptography
cryptography >= 2.6.1

# for mysql
mysql-connector-python >= 8.0.15